add components and profile page

+++ base/views.py
@@ -75,7 +75,14 @@ def home(request):
     
     topics = Topic.objects.all()
     room_count = rooms.count()
-    context = {'rooms': rooms, 'topics':topics, 'room_count':room_count}
+    # room_messages = Message.objects.all()
+    room_messages = Message.objects.filter(Q(room__topic__name__icontains=q))
+    
+    
+    context = {'rooms': rooms, 
+               'topics': topics, 
+               'room_count': room_count,
+               'room_messages': room_messages}
     return render(request, 'base/home.html', context)
 
 def room(request, pk):
@@ -96,6 +103,12 @@ def room(request, pk):
     context = {'room': room, 'room_messages': room_messages, 'participants':participants}
     
     return render(request, 'base/room.html', context)
+
+
+def userProfile(request, pk):
+    user = User.objects.get(id=pk)
+    context = {'user': user}
+    return render(request, 'base/profile.html', context )
 
 @login_required(login_url='login')
 def createRoom(request):
@@ -131,7 +144,21 @@ def updateRoom(request, pk):
 @login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id = pk)
+    if request.user !=room.host:
+        return HttpResponse('You are not allowed here!')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     return render(request, 'base/delete.html', {'obj':room})
+
+@login_required(login_url='login')
+def deleteMessage(request, pk):
+    message = Message.objects.get(id = pk)
+    if request.user !=message.user:
+        return HttpResponse('You are not allowed here!')
+    
+    if request.method == 'POST':
+        message.delete()
+        return redirect('home')
+    return render(request, 'base/delete.html', {'obj':message})
